Fixes bug 1367466 - Rename to Jansky
This renames the Processor project to Jansky which is easier to distinguish from
"Socorro processor" than "processor".
