+++ docs/conf.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Processor documentation build configuration file, created by
+# Jansky documentation build configuration file, created by
 # sphinx-quickstart on Wed Apr 12 11:20:13 2017.
 #
 # This file is execfile()d with the current directory set to its
@@ -45,7 +45,7 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'Processor'
+project = u'Jansky'
 copyright = u'2017, Socorro Team'
 author = u'Socorro Team'
 
@@ -99,7 +99,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'Processordoc'
+htmlhelp_basename = 'Janskydoc'
 
 
 # -- Options for LaTeX output ---------------------------------------------
@@ -126,7 +126,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'Processor.tex', u'Processor Documentation',
+    (master_doc, 'Jansky.tex', u'Jansky Documentation',
      u'Socorro Team', 'manual'),
 ]
 
@@ -136,7 +136,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'processor', u'Processor Documentation',
+    (master_doc, 'jansky', u'Jansky Documentation',
      [author], 1)
 ]
 
@@ -147,8 +147,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'Processor', u'Processor Documentation',
-     author, 'Processor', 'One line description of project.',
+    (master_doc, 'Jansky', u'Jansky Documentation',
+     author, 'Jansky', 'One line description of project.',
      'Miscellaneous'),
 ]
