+++ jansky/rules/mozilla_transform_rules.py
@@ -10,8 +10,8 @@ import time
 from sys import maxsize
 
 
-from processor.util import get_date_from_crash_id, datetime_from_isodate_string, utc_now
-from processor.rule import Rule
+from jansky.util import get_date_from_crash_id, datetime_from_isodate_string, utc_now
+from jansky.rule import Rule
 
 from urllib.parse import unquote_plus
