+++ setup.cfg
@@ -5,7 +5,7 @@ tag = True
 
 [bumpversion:file:setup.py]
 
-[bumpversion:file:processor/__init__.py]
+[bumpversion:file:jansky/__init__.py]
 
 [wheel]
 universal = 1
